refactor: credential check as ProofRequest method
- move check_credentials_against_proof_request into ProofRequest::check_credentials under requests/credential_check.rs
- replace manual constraint depth/node checks with validate_constraints()
- drop tests that duplicate world-id-primitives coverage
