+++ walletkit-core/Cargo.toml
@@ -26,6 +26,7 @@ backon = "1.6"
 base64 = { version = "0.22" }
 hex = "0.4"
 hkdf = "0.12"
+itertools = "0.14.0"
 log = "0.4"
 rand = "0.8"
 reqwest = { version = "0.12", default-features = false, features = ["json"] }
@@ -64,6 +65,7 @@ reqwest = { version = "0.12", default-features = false, features = [
 rustls = { version = "0.23", features = ["ring"] }
 
 [dev-dependencies]
+bon = "3"
 alloy = { version = "2", default-features = false, features = [
   "getrandom",
   "json",
