+++ walletkit-core/src/lib.rs
@@ -131,9 +131,6 @@ pub use user_agent::{UserAgent, UserAgentBuilder};
 /// Proof requests and responses in World ID v4.
 pub mod requests;
 
-/// Pre-flight check of whether stored credentials can satisfy a [`requests::ProofRequest`].
-pub mod proof_request_credential_constraints_check;
-
 mod proof;
 pub use proof::OwnershipProof;
