+++ walletkit-core/src/requests/mod.rs
@@ -1,3 +1,5 @@
+pub mod credential_check;
+
 use world_id_core::requests::{
     ProofRequest as CoreProofRequest, ProofResponse as CoreProofResponse,
 };
